Выпонит рефакторинг структуры папок проекта
