+++ src/pages/favorites-page/components/favorites-empty/favorites-empty.tsx
@@ -1,0 +1,12 @@
+function FavoriteEmpty(): JSX.Element {
+  return (
+    <div className="favorites__status-wrapper">
+      <b className="favorites__status">Nothing yet saved.</b>
+      <p className="favorites__status-description">
+        Save properties to narrow down search or plan your future trips.
+      </p>
+    </div>
+  );
+}
+
+export { FavoriteEmpty };
