+++ src/pages/main-page/components/locations-list/locations-list.tsx
