+++ src/pages/offer-page/components/gallery/gallery.tsx
@@ -1,4 +1,4 @@
-import { OfferType } from '../../types/types';
+import { OfferType } from '../../../../types/types';
 import GalleryItem from '../gallery-item/gallery-item';
 
 type GalleryProps = Pick<OfferType, 'images'>;
