+++ src/pages/offer-page/components/reviews-item/utils.ts
